host_app: move path definitions to platforms.js and unify check-configurtion
